+++ examples/dcloud-lab.sh
@@ -1,5 +1,0 @@
-export VMANAGE_IP='198.18.1.10'
-export VMANAGE_USER='admin'
-export VMANAGE_PASSWORD='admin'
-export CX_PID=0
-